refactor: Fix isMobile state
