+++ src/@types/GlobalNav.d.ts
@@ -1,6 +1,5 @@
 import { Dispatch, SetStateAction } from "react";
 
 export interface GlobalNavProps {
-  isMobile?: boolean;
   setShowSidebar?: Dispatch<SetStateAction<boolean>>;
 }
