+++ src/components/GlobalNav.tsx
@@ -1,14 +1,14 @@
+import { useRef, useState, useEffect } from "react";
 import { Link } from "react-router-dom";
 import useAuthStore from "../store/useAuthStore";
 import deleteCookie from "../utils/delete-cookie";
 import { GlobalNavProps } from "../@types/GlobalNav";
 import "./GlobalNav.css";
 
-export default function GlobalNav({
-  isMobile,
-  setShowSidebar,
-}: GlobalNavProps) {
+export default function GlobalNav({ setShowSidebar }: GlobalNavProps) {
+  const [isMobile, setIsMobile] = useState<boolean>(false);
   const { user, setUser } = useAuthStore();
+  const nav = useRef<HTMLDivElement>(null);
 
   const handleLogout = () => {
     localStorage.removeItem("access-token");
@@ -25,8 +25,28 @@ export default function GlobalNav({
     setShowSidebar((x) => !x);
   };
 
+  const handleResize = () => {
+    const { current } = nav;
+
+    if (!current) {
+      return;
+    }
+
+    setIsMobile(current.offsetWidth <= 660);
+  };
+
+  useEffect(() => {
+    handleResize();
+
+    window.addEventListener("resize", handleResize);
+
+    return () => {
+      window.removeEventListener("resize", handleResize);
+    };
+  }, []);
+
   return (
-    <div className="nav">
+    <div className="nav" ref={nav}>
       <div className="nav__items">
         <Link to="/" className="nav__item--head">
           Sinfroad
